requirements.txt
asgiref==3.8.1
Django==5.1.4
django-tastypie==0.15.0
python-dateutil==2.9.0.post0
python-mimeparse==2.0.0
six==1.17.0
sqlparse==0.5.3
